Move addressH to Deposit hierarchy, it depends on deposit library
